Fix shutdown, safety audit findings across sensor/dive_log/main
Shutdown fixes (main.c, screen_power_off.c, st7789v.h):
- Uncomment system_shutdown() call in screen_power_off_action()
- Fix backlight pin: PA2→PA9 (TFT_BLK/TIM2_CH3 on G474)
- Fix EXTI wake: remove wrong EXTI9_5/PA8, use BTN_M_Pin/EXTI2
- Add SLPIN+DISPOFF SPI commands before peripheral stop
- Refresh IWDG immediately before Stop1 for full ~6s window
- Stop ADC3/ADC4 DMA in shutdown (were left armed)
- Add ST7789_SLPIN/DISPOFF constants and SendCmd declaration

sensor.c safety fixes:
- Atomic sensor snapshot before multi-field reads (H-2)
- Narrowing cast on battery voltage multiply (H-9/10)

dive_log.c/h safety fixes:
- dive_log_erase(): check all sector erase returns, return bool (C-1)
- dive_log_iterate(): fix early-exit on entry_count not write_offset (H-1)
- dive_log_iterate(): fix unsigned wrap in reverse loop (H-4)
- dive_log_tick_1hz(): atomic sensor_data snapshot via critical section (H-2)
- recover_write_offset(): bound scan by entry_count to prevent ~52s init hang (M-2)
- sums[] moved from static to stack (non-reentrant static removed) (M-1)

main.c safety fixes:
- Check lv_disp_drv_register() return for NULL before use (H-1)
- Remove obsolete adc2_calculate_offsets_two_gain() prototype (M-3)
- Remove debug sentinel variables offset_counts_ch1/ch2 (L-1)
- Add static to emulate_profile/descend/depth globals (L-2)
- Add atomicity comment to MS5837 pressure struct writes (M-1)
- Remove SPI2_LoopbackTest() dead comment (L-3)

screen_dive_log.c:
- Fix lv_obj_set_pos() narrowing: explicit lv_coord_t casts (H-6)

CLAUDE.md:
- Add Type Safety Patterns section: narrowing, widening, memset/snprintf

+++ STM32CubeIDE/Application/User/Core/sensor.c
@@ -202,8 +202,8 @@ void sensor_data_update(sensor_id_t SensorID, uint16_t val) {
 		// Copy volatile reference to local to prevent torn read
 		uint32_t vref_uv = sensor_data.reference_uv;
 		//voltage_uv = adc_to_uV_at_sensor(val, vref_uv, 1);
-		voltage_uv =__LL_ADC_CALC_DATA_TO_VOLTAGE((int32_t)vref_uv/100, val, LL_ADC_RESOLUTION_12B);
-			sensor_data.battery_voltage_mv = voltage_uv * 3U; //(int32_t) ((voltage_uv / 100)*3)/2; //using voltage divider 200kO:100kO
+		voltage_uv = (uint32_t)__LL_ADC_CALC_DATA_TO_VOLTAGE((int32_t)vref_uv/100, val, LL_ADC_RESOLUTION_12B);
+			sensor_data.battery_voltage_mv = (int32_t)(voltage_uv * 3U); //(int32_t) ((voltage_uv / 100)*3)/2; //using voltage divider 200kO:100kO
 			if (sensor_data.battery_voltage_mv <= BAT_MIN_MV) {
 				sensor_data.battery_percentage = 0;
 			} else if (sensor_data.battery_voltage_mv >= BAT_MAX_MV) {
